pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tusk-cluster"
version = "0.1.0"
description = "Cluster mode plugin for Tusk - distributed queries with DataFusion"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
authors = [
    { name = "Jearel Alcantara" }
]
keywords = ["tusk", "plugin", "cluster", "datafusion", "distributed"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Web Environment",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Database",
]

dependencies = [
    "tuskdata>=0.2.0",
    "datafusion>=51.0.0",
    "pyarrow>=17.0.0",
    "psutil>=5.9",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
]

[project.entry-points."tusk.plugins"]
cluster = "tusk_cluster:ClusterPlugin"

[project.urls]
Homepage = "https://github.com/tuskdata/tusk-cluster"
Repository = "https://github.com/tuskdata/tusk-cluster"

[tool.hatch.build.targets.wheel]
packages = ["tusk_cluster"]
